+++ packages/spec/mock/config/simple-extends/config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  port: 8081,
+  https: true,
+};